Implements PvE and PvP deck buttons
In this commit a second button has been added for PvP decks. Clicking it
will generate a deck with different parameters, which thus far are
hard-coded.
In order to make this possible, the actual deck generation has been
wrapped into the functions gen_pve_deck() and gen_pvp_deck()
respectively. They are then called and the results are written using the
function write_deck(). The main() function now only contains two
buttons, each of which will call write_deck() with the parameters
according to the mode specified on the button.